Add pixel handler
